Move Provider proto and related kotlin code to a shared location.
These will be used inside both kingdom and duchy.
