+++ src/main/kotlin/org/wfanet/measurement/api/v2alpha/Principal.kt
@@ -12,11 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.measurement.kingdom.service.api.v2alpha
-
-import org.wfanet.measurement.api.v2alpha.DataProviderKey
-import org.wfanet.measurement.api.v2alpha.ModelProviderKey
-import org.wfanet.measurement.api.v2alpha.ResourceKey
+package org.wfanet.measurement.api.v2alpha
 
 /**
  * Identifies the sender of an inbound gRPC request.
