+++ src/main/kotlin/org/wfanet/measurement/api/v2alpha/PrincipalServerInterceptor.kt
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.measurement.kingdom.service.api.v2alpha
+package org.wfanet.measurement.api.v2alpha
 
 import com.google.protobuf.ByteString
 import io.grpc.BindableService
