+++ src/main/kotlin/org/wfanet/measurement/api/v2alpha/testing/BUILD.bazel
@@ -11,8 +11,8 @@ kt_jvm_library(
     name = "testing",
     srcs = ["MetadataPrincipalServerInterceptor.kt"],
     deps = [
+        "//src/main/kotlin/org/wfanet/measurement/api/v2alpha:principal_server_interceptor",
         "//src/main/kotlin/org/wfanet/measurement/common/identity",
-        "//src/main/kotlin/org/wfanet/measurement/kingdom/service/api/v2alpha:principal_server_interceptor",
         "@wfa_common_jvm//imports/java/org/junit",
         "@wfa_common_jvm//imports/kotlin/kotlinx/coroutines:core",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/identity",
