+++ src/main/kotlin/org/wfanet/measurement/common/identity/BUILD.bazel
@@ -33,3 +33,18 @@ kt_jvm_library(
         "@wfa_common_jvm//imports/java/io/grpc/stub",
     ],
 )
+
+kt_jvm_library(
+    name = "textproto_file_principal_lookup",
+    srcs = [
+        "TextprotoFilePrincipalLookup.kt",
+    ],
+    deps = [
+        "//src/main/kotlin/org/wfanet/measurement/api/v2alpha:principal",
+        "//src/main/kotlin/org/wfanet/measurement/api/v2alpha:principal_server_interceptor",
+        "//src/main/kotlin/org/wfanet/measurement/api/v2alpha:resource_key",
+        "//src/main/proto/wfa/measurement/config:authority_key_to_principal_map_kt_jvm_proto",
+        "@wfa_common_jvm//imports/java/com/google/protobuf",
+        "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common",
+    ],
+)
