+++ src/main/kotlin/org/wfanet/measurement/integration/common/BUILD.bazel
@@ -45,7 +45,7 @@ kt_jvm_library(
     ],
     visibility = ["//visibility:public"],
     deps = [
-        "//src/main/kotlin/org/wfanet/measurement/kingdom/service/api/v2alpha/testing",
+        "//src/main/kotlin/org/wfanet/measurement/api/v2alpha/testing",
         "//src/main/kotlin/org/wfanet/measurement/common/identity/testing",
         "//src/main/kotlin/org/wfanet/measurement/kingdom/deploy/common/service:data_services",
         "//src/main/kotlin/org/wfanet/measurement/kingdom/service/internal/testing/integration",
