+++ src/main/kotlin/org/wfanet/measurement/kingdom/deploy/gcloud/spanner/common/BUILD.bazel
@@ -10,7 +10,7 @@ kt_jvm_library(
     srcs = glob(["*.kt"]),
     deps = [
         "//src/main/kotlin/org/wfanet/measurement/common/identity",
-        "//src/main/proto/wfa/measurement/internal/kingdom:provider_kt_jvm_proto",
+        "//src/main/proto/wfa/measurement/internal/common:provider_kt_jvm_proto",
         "@wfa_common_jvm//imports/java/com/google/cloud/spanner",
         "@wfa_common_jvm//imports/java/com/google/protobuf",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common",
