+++ src/main/kotlin/org/wfanet/measurement/kingdom/service/internal/BUILD.bazel
@@ -9,7 +9,7 @@ kt_jvm_library(
     name = "internal",
     srcs = glob(["*.kt"]),
     deps = [
-        "//src/main/proto/wfa/measurement/internal/kingdom:provider_kt_jvm_proto",
+        "//src/main/proto/wfa/measurement/internal/common:provider_kt_jvm_proto",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/grpc",
         "@wfa_common_jvm//src/main/kotlin/org/wfanet/measurement/common/identity",
     ],
