+++ src/main/kotlin/org/wfanet/measurement/kingdom/service/internal/testing/Assertion.kt
@@ -20,13 +20,14 @@ import com.google.common.truth.extensions.proto.ProtoTruth.assertThat
 import java.time.LocalDate
 import java.time.ZoneOffset
 import org.wfanet.measurement.common.toProtoDate
+import org.wfanet.measurement.internal.common.Provider
+import org.wfanet.measurement.internal.common.provider
 import org.wfanet.measurement.internal.kingdom.Exchange
 import org.wfanet.measurement.internal.kingdom.ExchangeStep
 import org.wfanet.measurement.internal.kingdom.ExchangeStepAttempt
 import org.wfanet.measurement.internal.kingdom.ExchangeStepAttemptsGrpcKt.ExchangeStepAttemptsCoroutineImplBase
 import org.wfanet.measurement.internal.kingdom.ExchangeStepsGrpcKt.ExchangeStepsCoroutineImplBase
 import org.wfanet.measurement.internal.kingdom.ExchangesGrpcKt.ExchangesCoroutineImplBase
-import org.wfanet.measurement.internal.kingdom.Provider
 import org.wfanet.measurement.internal.kingdom.exchange
 import org.wfanet.measurement.internal.kingdom.exchangeDetails
 import org.wfanet.measurement.internal.kingdom.exchangeStep
@@ -35,7 +36,6 @@ import org.wfanet.measurement.internal.kingdom.exchangeStepAttemptDetails
 import org.wfanet.measurement.internal.kingdom.getExchangeRequest
 import org.wfanet.measurement.internal.kingdom.getExchangeStepAttemptRequest
 import org.wfanet.measurement.internal.kingdom.getExchangeStepRequest
-import org.wfanet.measurement.internal.kingdom.provider
 
 private const val EXTERNAL_RECURRING_EXCHANGE_ID = 222L
 private const val EXTERNAL_MODEL_PROVIDER_ID = 666L
