+++ src/main/proto/wfa/measurement/config/BUILD.bazel
@@ -4,9 +4,12 @@ load("@wfa_common_jvm//build:kt_jvm_proto_library.bzl", "kt_jvm_proto_library")
 
 package(default_visibility = ["//visibility:public"])
 
+IMPORT_PREFIX = "/src/main/proto"
+
 proto_library(
     name = "duchy_rpc_config_proto",
     srcs = ["duchy_rpc_config.proto"],
+    strip_import_prefix = IMPORT_PREFIX,
 )
 
 java_proto_library(
@@ -18,4 +21,21 @@ kt_jvm_proto_library(
     name = "duchy_rpc_config_kt_jvm_proto",
     srcs = [":duchy_rpc_config_proto"],
     deps = [":duchy_rpc_config_java_proto"],
+)
+
+proto_library(
+    name = "authority_key_to_principal_map_proto",
+    srcs = ["authority_key_to_principal_map.proto"],
+    strip_import_prefix = IMPORT_PREFIX,
+)
+
+java_proto_library(
+    name = "authority_key_to_principal_map_java_proto",
+    deps = [":authority_key_to_principal_map_proto"],
+)
+
+kt_jvm_proto_library(
+    name = "authority_key_to_principal_map_kt_jvm_proto",
+    srcs = [":authority_key_to_principal_map_proto"],
+    deps = [":authority_key_to_principal_map_java_proto"],
 )
