+++ src/main/proto/wfa/measurement/config/authority_key_to_principal_map.proto
@@ -14,9 +14,9 @@
 
 syntax = "proto3";
 
-package wfa.measurement.api.v2alpha;
+package wfa.measurement.config;
 
-option java_package = "org.wfanet.measurement.api.v2alpha";
+option java_package = "org.wfanet.measurement.config";
 option java_multiple_files = true;
 
 message AuthorityKeyToPrincipalMap {
