+++ src/main/proto/wfa/measurement/internal/common/BUILD.bazel
@@ -1,0 +1,24 @@
+load("@rules_proto//proto:defs.bzl", "proto_library")
+load("@rules_java//java:defs.bzl", "java_proto_library")
+load("@wfa_common_jvm//build:kt_jvm_proto_library.bzl", "kt_jvm_proto_library")
+
+package(default_visibility = ["//visibility:public"])
+
+IMPORT_PREFIX = "/src/main/proto"
+
+proto_library(
+    name = "provider_proto",
+    srcs = ["provider.proto"],
+    strip_import_prefix = IMPORT_PREFIX,
+)
+
+java_proto_library(
+    name = "provider_java_proto",
+    deps = [":provider_proto"],
+)
+
+kt_jvm_proto_library(
+    name = "provider_kt_jvm_proto",
+    srcs = [":provider_proto"],
+    deps = [":provider_java_proto"],
+)
