+++ src/main/proto/wfa/measurement/internal/kingdom/BUILD.bazel
@@ -241,7 +241,7 @@ proto_library(
     strip_import_prefix = IMPORT_PREFIX,
     deps = [
         ":exchange_proto",
-        ":provider_proto",
+        "//src/main/proto/wfa/measurement/internal/common:provider_proto",
         "@com_google_googleapis//google/type:date_proto",
     ],
 )
@@ -260,7 +260,7 @@ kt_jvm_proto_library(
 proto_and_java_proto_library(
     name = "exchange_step",
     deps = [
-        ":provider_proto",
+        "//src/main/proto/wfa/measurement/internal/common:provider_proto",
         "@com_google_googleapis//google/type:date_proto",
         "@com_google_protobuf//:timestamp_proto",
     ],
@@ -272,7 +272,7 @@ proto_library(
     strip_import_prefix = IMPORT_PREFIX,
     deps = [
         ":exchange_step_proto",
-        ":provider_proto",
+        "//src/main/proto/wfa/measurement/internal/common:provider_proto",
         "@com_google_googleapis//google/type:date_proto",
         "@com_google_protobuf//:timestamp_proto",
     ],
@@ -311,7 +311,7 @@ proto_library(
     deps = [
         ":exchange_step_attempt_details_proto",
         ":exchange_step_attempt_proto",
-        ":provider_proto",
+        "//src/main/proto/wfa/measurement/internal/common:provider_proto",
         "@com_google_googleapis//google/type:date_proto",
     ],
 )
@@ -483,5 +483,3 @@ kt_jvm_proto_library(
 )
 
 proto_and_java_proto_library(name = "exchange_workflow")
-
-proto_and_java_proto_library(name = "provider")
