+++ src/main/proto/wfa/measurement/internal/kingdom/all_protos.bzl
@@ -68,7 +68,7 @@ KINGDOM_INTERNAL_ENTITY_PROTOS = [
     "//src/main/proto/wfa/measurement/internal/kingdom:measurement_log_entry_kt_jvm_proto",
     "//src/main/proto/wfa/measurement/internal/kingdom:model_provider_kt_jvm_proto",
     "//src/main/proto/wfa/measurement/internal/kingdom:protocol_config_kt_jvm_proto",
-    "//src/main/proto/wfa/measurement/internal/kingdom:provider_kt_jvm_proto",
+    "//src/main/proto/wfa/measurement/internal/common:provider_kt_jvm_proto",
     "//src/main/proto/wfa/measurement/internal/kingdom:recurring_exchange_details_kt_jvm_proto",
     "//src/main/proto/wfa/measurement/internal/kingdom:recurring_exchange_kt_jvm_proto",
 ]
