+++ src/main/proto/wfa/measurement/internal/kingdom/exchange_step_attempts_service.proto
@@ -17,9 +17,9 @@ syntax = "proto3";
 package wfa.measurement.internal.kingdom;
 
 import "google/type/date.proto";
+import "wfa/measurement/internal/common/provider.proto";
 import "wfa/measurement/internal/kingdom/exchange_step_attempt.proto";
 import "wfa/measurement/internal/kingdom/exchange_step_attempt_details.proto";
-import "wfa/measurement/internal/kingdom/provider.proto";
 
 option java_package = "org.wfanet.measurement.internal.kingdom";
 option java_multiple_files = true;
@@ -94,7 +94,7 @@ message FinishExchangeStepAttemptRequest {
   repeated ExchangeStepAttemptDetails.DebugLog debug_log_entries = 6;
 
   // Denormalized party that executes this ExchangeStepAttempt.
-  Provider provider = 7;
+  wfa.measurement.internal.common.Provider provider = 7;
 }
 
 message GetExchangeStepAttemptRequest {
@@ -105,5 +105,5 @@ message GetExchangeStepAttemptRequest {
   int32 attempt_number = 4;
 
   // Denormalized party that executes this ExchangeStepAttempt.
-  Provider provider = 5;
+  wfa.measurement.internal.common.Provider provider = 5;
 }
