+++ src/main/proto/wfa/measurement/internal/kingdom/exchanges_service.proto
@@ -17,8 +17,8 @@ syntax = "proto3";
 package wfa.measurement.internal.kingdom;
 
 import "google/type/date.proto";
+import "wfa/measurement/internal/common/provider.proto";
 import "wfa/measurement/internal/kingdom/exchange.proto";
-import "wfa/measurement/internal/kingdom/provider.proto";
 
 option java_package = "org.wfanet.measurement.internal.kingdom";
 option java_multiple_files = true;
@@ -44,7 +44,7 @@ message GetExchangeRequest {
   // there is at most one `Exchange` for any given date -- this is part of the
   // `Exchange`'s primary key.
   google.type.Date date = 2;
-  Provider provider = 3;
+  wfa.measurement.internal.common.Provider provider = 3;
 }
 
 // Request for `/Exchanges.CreateExchange`.
