--- src/main/proto/wfa/measurement/internal/kingdom/provider.proto
@@ -1,33 +1,0 @@
-// Copyright 2021 The Cross-Media Measurement Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-syntax = "proto3";
-
-package wfa.measurement.internal.kingdom;
-
-option java_package = "org.wfanet.measurement.internal.kingdom";
-option java_multiple_files = true;
-
-message Provider {
-  // External identifier of the Provider.
-  fixed64 external_id = 1;
-
-  // Type of the Provider.
-  enum Type {
-    TYPE_UNSPECIFIED = 0;
-    MODEL_PROVIDER = 1;
-    DATA_PROVIDER = 2;
-  }
-  Type type = 2;
-}
