+++ src/test/kotlin/org/wfanet/measurement/kingdom/service/api/v2alpha/BUILD.bazel
@@ -88,10 +88,10 @@ kt_jvm_test(
         "//src/main/proto/wfa/measurement/api/v2alpha:exchange_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/api/v2alpha:exchanges_service_kt_jvm_grpc",
         "//src/main/proto/wfa/measurement/api/v2alpha:exchanges_service_kt_jvm_proto",
+        "//src/main/proto/wfa/measurement/internal/common:provider_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/internal/kingdom:exchange_details_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/internal/kingdom:exchanges_service_kt_jvm_grpc",
         "//src/main/proto/wfa/measurement/internal/kingdom:exchanges_service_kt_jvm_proto",
-        "//src/main/proto/wfa/measurement/internal/kingdom:provider_kt_jvm_proto",
         "@wfa_common_jvm//imports/java/com/google/common/truth",
         "@wfa_common_jvm//imports/java/com/google/common/truth/extensions/proto",
         "@wfa_common_jvm//imports/java/com/google/protobuf",
@@ -138,10 +138,10 @@ kt_jvm_test(
         "//src/main/proto/wfa/measurement/api/v2alpha:exchange_steps_service_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/api/v2alpha:exchange_workflow_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/api/v2alpha:model_provider_kt_jvm_proto",
+        "//src/main/proto/wfa/measurement/internal/common:provider_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/internal/kingdom:exchange_step_kt_jvm_proto",
         "//src/main/proto/wfa/measurement/internal/kingdom:exchange_steps_service_kt_jvm_grpc",
         "//src/main/proto/wfa/measurement/internal/kingdom:exchange_steps_service_kt_jvm_proto",
-        "//src/main/proto/wfa/measurement/internal/kingdom:provider_kt_jvm_proto",
         "@wfa_common_jvm//imports/java/com/google/common/truth",
         "@wfa_common_jvm//imports/java/com/google/common/truth/extensions/proto",
         "@wfa_common_jvm//imports/java/com/google/protobuf",
